GTK4: merge gui_gtk4_x11.c into gui_gtk4.c
The GTK4 backend has no X11 dependency, so the _x11 suffix was
misleading. Merge gui_gtk4.c (menus, scrollbars, dialogs) and
gui_gtk4_x11.c (main GUI) into a single gui_gtk4.c file.

Also merge proto/gui_gtk4_x11.pro into proto/gui_gtk4.pro and
update Makefile, proto.h, and Filelist accordingly.
