+++ Filelist
@@ -503,7 +503,6 @@ SRC_UNIX =	\
 		src/gui_gtk4.c \
 		src/gui_gtk4_f.c \
 		src/gui_gtk4_f.h \
-		src/gui_gtk4_x11.c \
 		src/gui_gtk_res.xml \
 		src/gui_motif.c \
 		src/gui_xmdlg.c \
@@ -534,7 +533,6 @@ SRC_UNIX =	\
 		src/proto/gui_gtk_x11.pro \
 		src/proto/gui_gtk_gresources.pro \
 		src/proto/gui_gtk4.pro \
-		src/proto/gui_gtk4_x11.pro \
 		src/proto/gui_motif.pro \
 		src/proto/gui_xmdlg.pro \
 		src/proto/gui_x11.pro \
