--- src/proto/gui_gtk4_x11.pro
@@ -1,70 +1,0 @@
-/* gui_gtk4_x11.c */
-void gui_mch_prepare(int *argc, char **argv);
-void gui_mch_free_all(void);
-int gui_mch_is_blinking(void);
-int gui_mch_is_blink_off(void);
-void gui_mch_set_blinking(long waittime, long on, long off);
-void gui_mch_stop_blink(int may_call_gui_update_cursor);
-void gui_mch_start_blink(void);
-int gui_mch_early_init_check(int give_message);
-int gui_mch_init_check(void);
-int gui_mch_init(void);
-void gui_mch_new_colors(void);
-int gui_mch_open(void);
-void gui_mch_exit(int rc);
-int gui_mch_get_winpos(int *x, int *y);
-void gui_mch_set_winpos(int x, int y);
-int gui_mch_maximized(void);
-void gui_mch_unmaximize(void);
-void gui_mch_newfont(void);
-void gui_mch_settitle(char_u *title, char_u *icon);
-void gui_mch_set_shellsize(int width, int height, int min_width, int min_height, int base_width, int base_height, int direction);
-void gui_mch_get_screen_dimensions(int *screen_w, int *screen_h);
-void gui_mch_enable_menu(int showit);
-void gui_mch_show_toolbar(int showit);
-void gui_mch_set_dark_theme(int dark);
-int gui_mch_adjust_charheight(void);
-char_u *gui_mch_font_dialog(char_u *oldval);
-int gui_mch_init_font(char_u *font_name, int fontset);
-GuiFont gui_mch_get_font(char_u *name, int report_error);
-char_u *gui_mch_get_fontname(GuiFont font, char_u *name);
-void gui_mch_free_font(GuiFont font);
-int gui_mch_expand_font(optexpand_T *args, int *numMatches, char_u ***matches);
-guicolor_T gui_mch_get_color(char_u *name);
-guicolor_T gui_mch_get_rgb_color(int r, int g, int b);
-void gui_mch_set_fg_color(guicolor_T color);
-void gui_mch_set_bg_color(guicolor_T color);
-void gui_mch_set_sp_color(guicolor_T color);
-guicolor_T gui_mch_get_rgb(guicolor_T pixel);
-void gui_mch_clear_block(int row1, int col1, int row2, int col2);
-void gui_mch_clear_all(void);
-void gui_mch_delete_lines(int row, int num_lines);
-void gui_mch_insert_lines(int row, int num_lines);
-void gui_mch_draw_hollow_cursor(guicolor_T color);
-void gui_mch_draw_part_cursor(int w, int h, guicolor_T color);
-void gui_mch_flash(int msec);
-void gui_mch_invert_rectangle(int r, int c, int nr, int nc);
-void gui_mch_update(void);
-int gui_mch_wait_for_chars(long wtime);
-void gui_mch_flush(void);
-void gui_mch_beep(void);
-void *gui_mch_get_display(void);
-void gui_mch_iconify(void);
-void gui_mch_set_foreground(void);
-void gui_mch_getmouse(int *x, int *y);
-void gui_mch_setmouse(int x, int y);
-void gui_mch_mousehide(int hide);
-int gui_mch_haskey(char_u *name);
-void gui_mch_forked(void);
-void gui_mch_enable_scrollbar(scrollbar_T *sb, int flag);
-void gui_mch_menu_grey(vimmenu_T *menu, int grey);
-void gui_mch_menu_hidden(vimmenu_T *menu, int hidden);
-void gui_mch_draw_menubar(void);
-void gui_mch_show_tabline(int showit);
-int gui_mch_showing_tabline(void);
-void gui_mch_update_tabline(void);
-void gui_mch_set_curtab(int nr);
-void gui_mch_drawsign(int row, int col, int typenr);
-void *gui_mch_register_sign(char_u *signfile);
-void gui_mch_destroy_sign(void *sign);
-int gui_gtk_draw_string(int row, int col, char_u *s, int len, int flags);
